Add authorization checks in FollowService, update UserDetails checks in DataInitializer
Modified the FollowService class to include references to PostService and UserService in order to implement authorization checks before showing the feed to a user. The service now ensures that the user is authenticated before allowing access to the feed. Updated FollowRepository to use "findAllByUserId" instead of "findAllByUser" to match the updated fields in Follow Entity. Also, in DataInitializer, added exception catching when attempting to load an existing user in the 'run' method to handle cases if the user does not presently exist.

Additionally, updated user-service's application.properties file to reset the database (DDL set to create from update) to accommodate the modified database schema.

Documentation (README.md) has also been significantly updated with crucial details about the project, including information about the microservices architectures used, modules description, technologies utilized and steps to run the project.